Modify ClaimsStrategy so that it only tries to call the authentication handler once.
Attached the user principle to HttpContext.

Had an issue where the Authentication was been called elsewhere and it caused an exception on the 2nd call.
Also, couldn't access the identity outside the ClaimStrategy function because it was been thrown away, changed to attach the principle to the httpContext so it can be accessed later.